requirements.txt
fastapi
uvicorn
transformers
torch
sentence-transformers
faiss-cpu
pydantic